chore: add test
