+++ pressbooks-multi-institution.php
@@ -18,19 +18,19 @@ use PressbooksMultiInstitution\Bootstrap;
 use PressbooksMultiInstitution\Database\Migration;
 
 // TODO: Check if this is the best way to check for Pressbooks.
-//if (!class_exists('Pressbooks\Book')) {
-if (file_exists(__DIR__ . '/vendor/autoload.php')) {
-    require_once __DIR__ . '/vendor/autoload.php';
-} else {
-    $title = __('Missing dependencies', 'PressbooksMultiInstitution');
-    $body = __(
-        'Please run <code>composer install</code> from the root of the plugin directory.',
-        'pressbooks-multi-institution'
-    );
+if (!class_exists('Pressbooks\Book')) {
+    if (file_exists(__DIR__ . '/vendor/autoload.php')) {
+        require_once __DIR__ . '/vendor/autoload.php';
+    } else {
+        $title = __('Missing dependencies', 'PressbooksMultiInstitution');
+        $body = __(
+            'Please run <code>composer install</code> from the root of the plugin directory.',
+            'pressbooks-multi-institution'
+        );
 
-    wp_die("<h1>{$title}</h1><p>{$body}</p>");
+        wp_die("<h1>{$title}</h1><p>{$body}</p>");
+    }
 }
-//}
 
 register_activation_hook(__FILE__, [Migration::class, 'migrate']);
 register_deactivation_hook(__FILE__, [Migration::class, 'rollback']);
