+++ tests/Feature/Actions/AssignBookToInstitutionTest.php
@@ -9,7 +9,6 @@ use PressbooksMultiInstitution\Models\Institution;
 use Tests\TestCase;
 use Tests\Traits\Assertions;
 use Tests\Traits\CreatesModels;
-use WP_Hook;
 
 class AssignBookToInstitutionTest extends TestCase
 {
@@ -179,34 +178,5 @@ class AssignBookToInstitutionTest extends TestCase
         );
 
         $this->assertDatabaseEmpty('institutions_blogs');
-    }
-
-    /**
-     * This method runs a callback without triggering a given filter.
-     * It returns the expected value from the callback.
-     *
-     * TODO: this method could be useful elsewhere, might be worth to extract it.
-     *
-     * @param  string  $hook The hook name to be skipped
-     * @param  callable  $callback The method that should run
-     * @return mixed The value from the callback method
-     */
-    protected function runWithoutFilter(string $hook, callable $callback): mixed
-    {
-        global $wp_filter;
-
-        $handler = $wp_filter[$hook] ?? null;
-
-        if ($handler instanceof WP_Hook) {
-            unset($wp_filter[$hook]);
-        }
-
-        $response = call_user_func($callback);
-
-        if ($handler instanceof WP_Hook) {
-            $wp_filter[$hook] = $handler;
-        }
-
-        return $response;
     }
 }
